Created website header with empty nav
