+++ src/Components/Navbar.css
@@ -1,0 +1,20 @@
+* {
+  padding: 0;
+  margin: 0;
+}
+
+.logo {
+  display: flex;
+  align-items: center;
+  margin-left: 2rem;
+}
+
+.logo h1 {
+  font-size: 2rem;
+}
+
+.logo-image {
+  height: 5rem;
+  padding-right: 3rem;
+}
+
