--- src/Pages/Missions.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function Missions() {
-  return <div>Missions</div>;
-}
-
-export default Missions;
