--- src/Pages/Profile.css
@@ -1,17 +1,0 @@
-* {
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
-}
-
-.profile-container {
-  padding: 0 4%;
-  display: grid;
-  gap: 5%;
-  grid-template-columns: 1fr 1fr;
-}
-
-.mission-lists,
-.rockets-list {
-  margin-bottom: 0.5rem;
-}
