--- src/Pages/Rockets.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function Rockets() {
-  return <div>Rockets</div>;
-}
-
-export default Rockets;
